Release 1 | Beta 0 | Development 3 (V0.2.6)
* Improved CDLODTerrain
* Improved accuracy of FPSLimiter

* Added convenience methods for setting the x, y and z
  positions of a transform/object separately
* Added support for more than one shadow map in a render
  scene batch
* Added an alternative constructor for FramebufferTexture
* Added batch system to RenderScene to avoid repeatedly
  assigning the same uniforms

* Changed getDelta() methods to be of type float to be
  more accurate
* Changed debugging settings to debug

* Fixed bug: Frustum culling did not take into account the
             scale of the object
* Fixed bug: Shadow's would repeat themselves if far enough
             away
* Fixed bug: Terrain normals were incorrectly calculated
* Fixed bug: DebugCamera was incorrectly calculating mouse
             movement
* Fixed bug: GUIDropDownMenu and hence GUIDropDownList did
             not assign added button's transform
* Fixed bug: Horizontal GUISlider did not work correctly

+++ UnnamedEngine/Source/utils/FPSUtils.cpp
@@ -41,7 +41,7 @@ void FPSCalculator::update() {
 	//Only calculate everything if necessary
 	if (mode != OFF) {
 		//Obtain the current time in milliseconds
-		long current = getMilliseconds();
+		float current = getMilliseconds();
 
 		//With a high FPS the delta will become 0, but to stop this causing issues
 		//the current delta will only be assigned when this is not the case and
@@ -97,7 +97,7 @@ void FPSLimiter::setMaxFPS(unsigned int maxFPS) {
 	//Prevent a divide by 0
 	if (maxFPS != 0)
 		//Calculate the target delta to achieve the desired frame rate
-		targetDelta = (long) (1000.0f / (float) maxFPS);
+		targetDelta = (1000.0f / (float) maxFPS);
 }
 
 void FPSLimiter::startFrame() {
@@ -109,7 +109,7 @@ void FPSLimiter::endFrame() {
 	//Ensure there is a frame cap
 	if (maxFPS > 0) {
 		//Calculate the difference in the target delta, and the current one
-		long difference = targetDelta - (getMilliseconds() - startOfFrame);
+		float difference = targetDelta - (getMilliseconds() - startOfFrame);
 
 		//Pause the thread for the time necessary
 		if (difference > 0)
